requirements.txt.txt
streamlit
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
google-generativeai
sentence-transformers